% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHaploStates.R
\name{allStatus}
\alias{allStatus}
\title{Generates all binary strings with n 0s and m 1s.}
\usage{
allStatus(n, m)
}
\arguments{
\item{n}{The number of 0s.}

\item{m}{The number of 1s.}
}
\description{
Generates all binary strings with n 0s and m 1s.
}
\author{
Maryam Ghareghani
}

