% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeCellTypesFormat}
\alias{changeCellTypesFormat}
\title{TODO Write function description or remove from the file!!!}
\usage{
changeCellTypesFormat(stateFile)
}
\arguments{
\item{stateFile}{TODO ...}
}
\description{
TODO Write function description or remove from the file!!!
}
\author{
Maryam Ghareghani
}

