% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeNBparamsFormat}
\alias{changeNBparamsFormat}
\title{TODO Write function description or remove from the file!!!}
\usage{
changeNBparamsFormat(infoFile, K)
}
\arguments{
\item{infoFile}{TODO ...}

\item{K}{The number of chromosomes (autosomes).}
}
\description{
TODO Write function description or remove from the file!!!
}
\author{
Maryam Ghareghani
}

