% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeRCformat}
\alias{changeRCformat}
\title{Changes the format of all files to the format used in the SV calling wrapper function.}
\usage{
changeRCformat(RCfile, outputDir, bamNamesFile = "bamNames.txt")
}
\arguments{
\item{RCfile}{TODO ...}

\item{outputDir}{The directory containing the output files.}

\item{bamNamesFile}{TODO ...}
}
\description{
Changes the format of all files to the format used in the SV calling wrapper function.
}
\author{
Maryam Ghareghani
}

