% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrNum.R
\name{chrNumber}
\alias{chrNumber}
\title{returns the number of a chromosome, given the input string}
\usage{
chrNumber(chr)
}
\arguments{
\item{chr}{a string like "chri"}
}
\description{
returns the number of a chromosome, given the input string
}
\author{
Maryam Ghareghani
}

