% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDispParAndSegType.R
\name{dispersionPar}
\alias{dispersionPar}
\title{Compute the dispersion paramters, given a segment type.}
\usage{
dispersionPar(segType, r, segLen = binLen, binLen, alpha = 0.05)
}
\arguments{
\item{segType}{A string containing a number of w characters followed by a number of c characters.}

\item{r}{Dispersion parameter for the total number of read counts (W+C in all cells) in a bin with a diploid copy number (2).}

\item{segLen}{The length of the segment.}

\item{binLen}{The length of the bins used for estimating the NB parameters.}

\item{alpha}{The coefficient of NB dispersion parameter for the number of background reads.}
}
\description{
Compute the dispersion paramters, given a segment type.
}
\author{
Maryam Ghareghani
}

