% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFitting.R
\name{estimateP}
\alias{estimateP}
\title{Estimates p parameter and saves the empirical mean-var plot in the \code{directory}.}
\usage{
estimateP(counts, directory = NULL)
}
\arguments{
\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}

\item{directory}{Directory containing the bam files.}
}
\description{
Estimates p parameter and saves the empirical mean-var plot in the \code{directory}.
}
\author{
Maryam Ghareghani
}

