% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFitting.R
\name{estimateR}
\alias{estimateR}
\title{Estimates r parameters}
\usage{
estimateR(counts, p)
}
\arguments{
\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}

\item{p}{The p parameter of the NB distribution.}
}
\description{
Estimates r parameters
}
\author{
Maryam Ghareghani
}

