% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filt}
\alias{filt}
\title{Filters the list of data.frames based on the list od indexes.}
\usage{
filt(list.df, list.idx)
}
\arguments{
\item{list.df}{A \code{list} of data.frames.}

\item{list.idx}{A \code{list} of indexes for subsetting rows of the \code{data.frame}.}
}
\description{
Filters the list of data.frames based on the list od indexes.
}
\author{
Maryam Ghareghani
}

