% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{genotype_class}
\alias{genotype_class}
\title{returns the genotyoe class (CN_loss, CN_gain, normal_CN2, or inv_CN2)}
\usage{
genotype_class(genotype.name)
}
\arguments{
\item{genotype.name}{The genotype or haplotype coding}
}
\description{
returns the genotyoe class (CN_loss, CN_gain, normal_CN2, or inv_CN2)
}
\author{
Maryam Ghareghani
}
