% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{get_GT_class_prob_table}
\alias{get_GT_class_prob_table}
\title{returns the genotyoe class (CN_loss, CN_gain, normal_CN2, or inv_CN2) probability dataframe}
\usage{
get_GT_class_prob_table(prob.tab)
}
\arguments{
\item{prob.tab}{The genotype or haplotype probability dataframe}
}
\description{
returns the genotyoe class (CN_loss, CN_gain, normal_CN2, or inv_CN2) probability dataframe
}
\author{
Maryam Ghareghani
}
