% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPossibleCNs.R
\name{getCNprob}
\alias{getCNprob}
\title{Generate a CN probability vector.}
\usage{
getCNprob(readCount, p, r, segLen, binLen, maxCN = 5, alpha = 0.05)
}
\arguments{
\item{readCount}{Total number of read counts (W+C) in all cells in this segment.}

\item{p}{The p parameter of the NB distribution.}

\item{r}{Dispersion parameter for the total number of read counts (W+C in all cells) in a bin with a diploid copy number (2).}

\item{segLen}{The length of the segment.}

\item{binLen}{The length of the bins used for estimating the NB parameters.}

\item{maxCN}{The maximum number of CNs.}

\item{alpha}{The coefficient of NB dispersion parameter for the number of background reads.}
}
\description{
Generate a CN probability vector.
}
\author{
Maryam Ghareghani
}

