% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{getGenotypeProbDataFrame}
\alias{getGenotypeProbDataFrame}
\title{Takes as input a haplotype probTable and outout the genotype probTable.}
\usage{
getGenotypeProbDataFrame(hapProbDF)
}
\arguments{
\item{hapProbDF}{A \code{data.frame} containing haplotype probability table.}
}
\description{
Takes as input a haplotype probTable and outout the genotype probTable.
}
\author{
Maryam Ghareghani
}

