% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{getGenotypeProbTable}
\alias{getGenotypeProbTable}
\title{Compute a noramlInvCN probability table given the haplotype probability table.}
\usage{
getGenotypeProbTable(haplotypeProbTable)
}
\arguments{
\item{haplotypeProbTable}{A haplotype probability table with the set of decoded (non-binary) haplotypes for rows and cells for columns.}
}
\description{
Compute a noramlInvCN probability table given the haplotype probability table.
}
\author{
Maryam Ghareghani
}

