% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHaploStates.R
\name{getGenotypeStatus}
\alias{getGenotypeStatus}
\title{Computes the total normal CN and inverted CN status corresponding to the input decoded (non-binary) haplotype status.}
\usage{
getGenotypeStatus(haplotypeStatus)
}
\arguments{
\item{haplotyeStatus}{A decoded (non-binary) haplotype status.}
}
\description{
Note that it doesn't compute the genotype (distribution of the CNs in the haplotypes).
}
\author{
Maryam Ghareghani
}

