% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R
\name{getlegend}
\alias{getlegend}
\title{helper function for heatmap}
\usage{
getlegend(a.gplot)
}
\arguments{
\item{a.gplot}{TODO write document}
}
\description{
helper function for heatmap
}
\author{
David Porubsky
}
