% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHaploStates.R
\name{getNextState}
\alias{getNextState}
\title{Computes the next binary status of a given status with equal number of 0s and 1s.}
\usage{
getNextState(currentState)
}
\arguments{
\item{currentState}{The current binary status.}
}
\description{
Computes the next binary status of a given status with equal number of 0s and 1s.
}
\author{
Maryam Ghareghani
}

