% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPossibleCNs.R
\name{getPossibleCNs}
\alias{getPossibleCNs}
\title{Returns all CNs with maximum probability.}
\usage{
getPossibleCNs(segCounts, p, chrCellsDispPars, binLen, alpha)
}
\arguments{
\item{segCounts}{W and C counts of one segment (a count \code{data.frame} with one row).}

\item{p}{The p parameter of the NB distribution.}

\item{chrCellsDispPars}{The dispersion parameters of cells for the chromosome where the segment resides.}

\item{binLen}{The length of the bins used for estimating the NB parameters.}

\item{alpha}{The coefficient of NB dispersion parameter for the number of background reads.}
}
\description{
Returns all CNs with maximum probability.
}
\author{
Maryam Ghareghani
}

