% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{getSegReadCounts}
\alias{getSegReadCounts}
\title{TODO Write function description or remove from the file!!!}
\usage{
getSegReadCounts(binRC, breakpointsFile, K, bin.size)
}
\arguments{
\item{binRC}{TODO ...}

\item{breakpointsFile}{TODO ...}

\item{K}{The number of chromosomes (autosomes).}

\item{bin.size}{The size of the bins.}
}
\description{
TODO Write function description or remove from the file!!!
}
\author{
Maryam Ghareghani
}

