% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHaploStates.R
\name{initialState}
\alias{initialState}
\title{Returns the first binary string with n 01 and m 1s.}
\usage{
initialState(n, m)
}
\arguments{
\item{n}{The number of 0s.}

\item{m}{The number of 1s.}
}
\description{
Returns the first binary string with n 01 and m 1s.
}
\author{
Maryam Ghareghani
}

