% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{jumpProbs}
\alias{jumpProbs}
\title{For each segment, computes the probabilities of having the same haplotype for all single cells}
\usage{
jumpProbs(probTable.l)
}
\arguments{
\item{probTable.l}{The list of segments probability table}
}
\description{
For each segment, computes the probabilities of having the same haplotype for all single cells
}
\author{
Maryam Ghareghani
}
