% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{NBfitplots}
\alias{NBfitplots}
\title{Create seperate PDF files for different chromosomes and in each file, create plots for W, C, and total read coverage in all cells with the fitted NB distribution.}
\usage{
NBfitplots(directory, counts, cellTypes, p, dispersion, bin.size)
}
\arguments{
\item{directory}{Directory containing the bam files.}

\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}

\item{cellTypes}{TODO ...}

\item{p}{The p parameter of the NB distribution.}

\item{dispersion}{The estimated dispersion parameters.}

\item{bin.size}{An \code{integer} (e.g. 2000) used to bin data for plotting}
}
\description{
Create seperate PDF files for different chromosomes and in each file, create plots for W, C, and total read coverage in all cells with the fitted NB distribution.
}
\author{
Maryam Ghareghani
}

