% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{newgetCellStatProbabilities}
\alias{newgetCellStatProbabilities}
\title{Returns a matrix of haplotype (rows) pobabilities in different single cells (columns).}
\usage{
newgetCellStatProbabilities(hapStatus, counts, chrCellTypes, p,
  chrCellsDispPars, binLen, alpha, haplotypeMode = FALSE)
}
\arguments{
\item{hapStatus}{A \code{vector} of decoded halotype status (strings).}

\item{counts}{A one row \code{data.frame} containing the read counts of the segment.}

\item{chrCellTypes}{Cell types in the chromosome where the segment lies.}

\item{p}{The p parameter of the NB distribution.}

\item{chrCellsDispPars}{The dispersion parameters of cells for the chromosome where the segment resides.}

\item{binLen}{The length of the bins used for estimating the NB parameters.}

\item{alpha}{The coefficient of NB dispersion parameter for the number of background reads.}

\item{haplotypeMode}{TODO ...}
}
\description{
Returns a matrix of haplotype (rows) pobabilities in different single cells (columns).
}
\author{
Maryam Ghareghani
}

