% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{newgetGenotypeProbTable}
\alias{newgetGenotypeProbTable}
\title{Compute a genotype probability table given the haplotype probability table.}
\usage{
newgetGenotypeProbTable(haplotypeProbTable)
}
\arguments{
\item{haplotypeProbTable}{A haplotype probability table with the set of decoded (non-binary) haplotypes for rows and cells for columns.}
}
\description{
Compute a genotype probability table given the haplotype probability table.
}
\author{
Maryam Ghareghani
}

