% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVcalling.R
\name{newSVcalling}
\alias{newSVcalling}
\title{Takes as an input an aggregate prob table and call each segment by the haplotype of the highest probability.}
\usage{
newSVcalling(aggProbTable, haplotypeMode = FALSE)
}
\arguments{
\item{aggProbTable}{A \code{matrix} containing the logarithm of aggregate probabilities of haplotype states in segments, (rows correspond to segments and columns correspond to haplotype state).}

\item{haplotypeMode}{TODO ...}
}
\description{
Takes as an input an aggregate prob table and call each segment by the haplotype of the highest probability.
}
\author{
Maryam Ghareghani
}

