% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{nonzero.cov.bins}
\alias{nonzero.cov.bins}
\title{Outputs the list of indices of the bins with non zero coverage.}
\usage{
nonzero.cov.bins(counts)
}
\arguments{
\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}
}
\description{
Outputs the list of indices of the bins with non zero coverage.
}
\author{
Maryam Ghareghani
}

