% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{normalizeProbTable}
\alias{normalizeProbTable}
\title{Normalize a probability table to a table in which each column sums up to 1.}
\usage{
normalizeProbTable(probTable)
}
\arguments{
\item{probTable}{A probability table.}
}
\description{
Normalize a probability table to a table in which each column sums up to 1.
}
\author{
Maryam Ghareghani
}

