% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBams.R
\name{readBams}
\alias{readBams}
\title{Read bam files from a directory and output them (the first mates only) as a list of \code{\link[GenomicRanges]{GRanges}} objects. It also writes the name of the bam files in a file.}
\usage{
readBams(directory, bamFilenames = "", unq = TRUE)
}
\arguments{
\item{directory}{Directory containing the bam files.}

\item{bamFilenames}{Name of a file to write the names of the bam files in.}

\item{unq}{Only the unique alignments are given as output, if unq = TRUE.}
}
\description{
Read bam files from a directory and output them (the first mates only) as a list of \code{\link[GenomicRanges]{GRanges}} objects. It also writes the name of the bam files in a file.
}
\author{
Maryam Ghareghani
}

