% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapAndGTprobsComputing.R
\name{regularizeProbTable}
\alias{regularizeProbTable}
\title{We assume that the posteriori probability for different status is uniform with probability regFactor and a probability computed based on the NB model with probability (1-regFactor).}
\usage{
regularizeProbTable(probTable, regFactor = 1e-10)
}
\arguments{
\item{probTable}{A non-regularized probability table in which every row corresponds to a status and each column corresponds to a single cell.}

\item{regFactor}{The regularization factor.}
}
\description{
Regularize the probability table according to the mentioned assumption.
}
\author{
Maryam Ghareghani
}

