% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R
\name{setPanelHeights}
\alias{setPanelHeights}
\title{helper function for heatmap}
\usage{
setPanelHeights(g, heights)
}
\arguments{
\item{g}{TODO write document}

\item{heights}{TODO write document}
}
\description{
helper function for heatmap
}
\author{
David Porubsky
}
