% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHaploStates.R
\name{sisterHaplotype}
\alias{sisterHaplotype}
\title{Takes as an input a haplotype state and returns the opposite haplotype state which has the same genotype.}
\usage{
sisterHaplotype(hapState)
}
\arguments{
\item{hapState}{A decoded haplotype state.}
}
\description{
Takes as an input a haplotype state and returns the opposite haplotype state which has the same genotype.
}
\author{
Maryam Ghareghani
}

