% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCcount.R
\name{splitChromosomes}
\alias{splitChromosomes}
\title{Splits a \code{data.frame} of read counts based on chromosomes and returns a \code{list} containing the splited dataframes.}
\usage{
splitChromosomes(counts)
}
\arguments{
\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}
}
\description{
Splits a \code{data.frame} of read counts based on chromosomes and returns a \code{list} containing the splited dataframes.
}
\author{
Maryam Ghareghani
}

