% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_SVcalling.R
\name{SVcalling.wrapper.func}
\alias{SVcalling.wrapper.func}
\title{Main function for the \pkg{\link{MaRyam}} package.}
\usage{
SVcalling.wrapper.func(bin.size, K, maximumCN, segmentsCounts, r, p, cellTypes,
  outputDir, hapMode)
}
\arguments{
\item{bin.size}{the size of the bins}

\item{K}{The number of chromosomes (autosomes). #TODO I suggest that user can define what chromosome wants to analyze. So this parameter should take vector of chromosome IDs => c(1:22,'X') or paste0('chr', c(1:22,'X'))}

\item{maximumCN}{Maximum CN in the segments for SV calling. #TODO make sure parameters correspond to maxCN paramter in getCNprob}

\item{segmentsCounts}{TODO ...}

\item{r}{TODO ...}

\item{p}{TODO ...}

\item{cellTypes}{TODO ... In function getSegType there is a parameter cellType. Are these paramters expecting the same thing? If not please use bit more informative name.}

\item{outputDir}{The outputDir containing the input and output files.}

\item{hapMode}{TODO ... #There is a haplotypeMode parameter in newSVcalling function. I guess they point to the same parameter.}
}
\description{
Wrapper function to run SV calling after having the NB parmeters and segment read counts.
}
\author{
Maryam Ghareghani
}

