% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_SVcalling.R
\name{SVcalling_wrapperFunc}
\alias{SVcalling_wrapperFunc}
\title{Main function for the \pkg{\link{MaRyam}} package.}
\usage{
SVcalling_wrapperFunc(bin.size, K, maximumCN, segmentsCounts, r, p, cellTypes,
  outputDir, haplotypeMode)
}
\arguments{
\item{bin.size}{the size of the bins}

\item{K}{The number of chromosomes (autosomes). #TODO I suggest that user can define what chromosome wants to analyze. So this parameter should take vector of chromosome IDs => c(1:22,'X') or paste0('chr', c(1:22,'X'))}

\item{maximumCN}{Maximum CN in the segments for SV calling. #TODO make sure parameters correspond to maxCN paramter in getCNprob}

\item{segmentsCounts}{The dataframe containing the segments chr and coordinates and W and C read counts over all cells}

\item{r}{The matrix of dispersion parameters of chromosomes(rows) and cells(columns) for CN=2 for a segment of size bin.size}

\item{p}{The p parameter of the NB distribution}

\item{cellTypes}{The matrix of cell types of chromosomes(rows) and cells(columns) TODO ... In function getSegType there is a parameter cellType. Are these paramters expecting the same thing? If not please use bit more informative name.}

\item{outputDir}{The outputDir containing the input and output files.}

\item{haplotypeMode}{a binary parameter showing wether we have the haplotype of WC cell types TODO ... #There is a haplotypeMode parameter in newSVcalling function. I guess they point to the same parameter.}
}
\description{
Wrapper function to run SV calling after having the NB parmeters and segment read counts.
}
\author{
Maryam Ghareghani
}
