% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{uniqueMappableBins}
\alias{uniqueMappableBins}
\title{Outputs the list of indices of the bins having the fraction of low mapq reads less than alpha.}
\usage{
uniqueMappableBins(counts, unique.counts, alpha = 0.05)
}
\arguments{
\item{counts}{A \code{data.frame} containing W and C read counts in some segments.}

\item{alpha}{The coefficient of NB dispersion parameter for the number of background reads.}

\item{uniqe.counts}{A \code{data.frame} containing W and C unique read counts in segments.}
}
\description{
Outputs the list of indices of the bins having the fraction of low mapq reads less than alpha.
}
\author{
Maryam Ghareghani
}

