% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCcount.R
\name{WCreadCounts}
\alias{WCreadCounts}
\title{Count the number of Watson and Crick reads and return a \code{dataframe} of read counts.}
\usage{
WCreadCounts(segments, cells.alignments)
}
\arguments{
\item{segments}{A list \code{\link[GenomicRanges]{GRanges}} objects (genomic intervals).}

\item{cells.alignmemts}{A \code{list} of \code{\link[GenomicRanges]{GRanges}} objects with Strand-specific read data. #TODO Is the object a GRangesList or a normal list of GRanges??? If is is a GRangesList please use \code{\link[GenomicRanges]{GRangesList}}}
}
\description{
Count the number of Watson and Crick reads and return a \code{dataframe} of read counts.
}
\author{
Maryam Ghareghani
}

