% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpProbs.R
\name{addJumpProbs}
\alias{addJumpProbs}
\title{Computes the probabilities of having the same genotype/haplotype for all single cells (jump probabilities)
returns a list containing the lists of probability tables of segments for each chromosome
In the output probability tables, the input and output jump probabilities are included for internal (not first and last) segments}
\usage{
addJumpProbs(probTable, simplified = TRUE)
}
\arguments{
\item{probTable.l.chroms}{}
}
\description{
Computes the probabilities of having the same genotype/haplotype for all single cells (jump probabilities)
returns a list containing the lists of probability tables of segments for each chromosome
In the output probability tables, the input and output jump probabilities are included for internal (not first and last) segments
}
\author{
Maryam Ghareghani
}
