% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeCellTypesFormat}
\alias{changeCellTypesFormat}
\title{Changes the format of the cell types file and gives as output the cell types matrix}
\usage{
changeCellTypesFormat(stateFile)
}
\arguments{
\item{stateFile}{The name of the file containing the cell types}
}
\description{
Changes the format of the cell types file and gives as output the cell types matrix
}
\author{
Maryam Ghareghani
}
