% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeNBparamsFormat}
\alias{changeNBparamsFormat}
\title{Changes the format of the NB parameters and gives as output p parameter and a matrix containing r parameters}
\usage{
changeNBparamsFormat(infoFile, K)
}
\arguments{
\item{infoFile}{The name of the file containing the NB parameters}

\item{K}{The number of chromosomes (autosomes).}
}
\description{
Changes the format of the NB parameters and gives as output p parameter and a matrix containing r parameters
}
\author{
Maryam Ghareghani
}
