% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{changeRCformat}
\alias{changeRCformat}
\title{Changes the format of all files to the format used in the SV calling wrapper function.}
\usage{
changeRCformat(RCfile, outputDir, bamNamesFile = "bamNames.txt")
}
\arguments{
\item{RCfile}{The name of the file containing bin read counts}

\item{outputDir}{The directory containing the output files.}

\item{bamNamesFile}{Outputs the names of the bam files in the order they are used in the tool}
}
\description{
Changes the format of all files to the format used in the SV calling wrapper function.
}
\author{
Maryam Ghareghani
}
