% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_formatChange.R
\name{getSegReadCounts}
\alias{getSegReadCounts}
\title{outputs the segment counts}
\usage{
getSegReadCounts(binRC, breakpointsFile, K, bin.size)
}
\arguments{
\item{binRC}{bin read counts splitted by chromosomes}

\item{breakpointsFile}{The name of the breakpoint file}

\item{K}{The number of chromosomes (autosomes).}

\item{bin.size}{The size of the bins.}
}
\description{
outputs the segment counts
}
\author{
Maryam Ghareghani, Sascha Meiers
}
