% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R
\name{plotHeatmapSegment}
\alias{plotHeatmapSegment}
\title{Main heatmap function}
\usage{
plotHeatmapSegment(dataFrame, plot.log = FALSE, file = NULL, aggProbs = F,
  CNV = 3)
}
\arguments{
\item{dataFrame}{The probability table for a segment with single cells in rows and genotypes/haplotypes in columns}

\item{plot.log}{Binary option for plotting the heatmap in log scale}

\item{file}{TODO write document}

\item{aggProbs}{TODO write document}

\item{CNV}{The maximum CN for plotting}
}
\description{
Main heatmap function
}
\author{
David Porubsky, Maryam Ghareghani
}
