% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R, R/plotting.R
\name{plotHeatmapSegment}
\alias{plotHeatmapSegment}
\alias{plotHeatmapSegment}
\title{Main heatmap function}
\usage{
plotHeatmapSegment(dataFrame, plot.log = FALSE, file = NULL,
  aggProbs = FALSE, CNV = 3)

plotHeatmapSegment(dataFrame, plot.log = FALSE, file = NULL,
  aggProbs = FALSE, CNV = 3)
}
\arguments{
\item{dataFrame}{The probability table for a segment with single cells in rows and genotypes/haplotypes in columns}

\item{plot.log}{Binary option for plotting the heatmap in log scale}

\item{file}{TODO write document}

\item{aggProbs}{TODO write document}

\item{CNV}{The maximum CN for plotting}

\item{dataFrame}{A \code{\link{data.frame}} object that containing genotype probabilities.}

\item{plot.log}{A logical indicating whether or not to plot in logarithmic scale.}

\item{file}{A file to export the plot to.}

\item{aggProbs}{A logical indicating whether or not to plot aggregate probability values.}

\item{CNV}{A copy number value until which the probability values are plotted.}
}
\description{
Main heatmap function

Plot probability table as a heatmap.
}
\author{
David Porubsky, Maryam Ghareghani

David Porubsky
}
