% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_plots.R
\name{plotHeatmapSegment}
\alias{plotHeatmapSegment}
\title{Main heatmap function}
\usage{
plotHeatmapSegment(dataFrame, ord.feature = NULL, plot.log = FALSE,
  file = NULL, aggProbs = F, CNV = 3, simplified = TRUE)
}
\arguments{
\item{dataFrame}{The probability table for a segment with single cells in rows and genotypes/haplotypes in columns}

\item{ord.feature}{The ordering feature used for sorting single cells (rows) in the heatmap. 
It can be NULL (kepping the input order), "SV_probs" (baded hierarchical clustering of SV_probs), "input_jump_p", and "output_jump_p".}

\item{plot.log}{Binary option for plotting the heatmap in log scale}

\item{file}{TODO write document}

\item{aggProbs}{TODO write document}

\item{CNV}{The maximum CN for plotting}

\item{simplified}{If to plot simplified probability table e.g ("CN_loss","inv_CN2","normal_CN2","CN_gain")}
}
\description{
Main heatmap function
}
\author{
David Porubsky, Maryam Ghareghani
}
